# Faster __new__/from_gdal